pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fixnow"
version = "0.1.0"
description = "Marketplace de serviços domésticos com segurança e rastreamento em tempo real"
authors = [
    { name = "Lucas Costa", email = "lucas@fixnow.com.br" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"

dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",

    # Banco de dados
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Validação
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.2.0",

    # Autenticação
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.1.0,<5.0",
    "python-multipart>=0.0.9",

    # Cache
    "redis>=5.0.0",

    # Variáveis de ambiente
    "python-dotenv>=1.0.0",

    # HTTP client
    "httpx>=0.27.0",

    # Kafka
    "aiokafka>=0.10.0",

    # Geolocalização
    "geopy>=2.4.1",
]

[project.optional-dependencies]
dev = [
    # Testes
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",

    # Qualidade de código
    "ruff>=0.4.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
